Management API is now Python 2.5 compatible.
Cert and key for management API can now be in separate files, default
behaviour assumes they are together but if management_key_path is passed
when instantiating ServiceManagementEndpoints, then that key will be
passed through to the underlying HTTPSConnection and hence SSL socket.
This is *required* under Python 2.5 as the httplib.HTTPSConnection
therein expects both to be set.

Fixed error handling code to work with lxml. Unfortunately there doesn't
seem to be a reliable way to use isinstance against an lxml ElementTree
instance due to the ElementTree class not being part of the public API.
The _get_wa_error method now accepts an extra argument for pre-parsed
response data in order to avoid type testing.